Move function from chat to speech_recognition
